Adding random salt in front of each notification string
This ensures the b64-encoded string we send over the wire won't always look the same.
If we always start with the <header_tag>, and the password hash doesn't change,
the strings we send over the wire always start with the same few characters. By adding
a random salt in front, we generate less consistent strings